+++ phase2/Game/app/src/main/java/com/example/game/Coin.java
@@ -5,13 +5,13 @@ import android.graphics.Color;
 
 class Coin extends GameObject {
 
-  Coin(float x, float y) {
+  Coin(double x, double y) {
     super(x, y);
     paint.setColor(Color.YELLOW);
   }
 
   void draw(Canvas canvas) {
-    canvas.drawCircle(getX(), getY(), 15, paint);
+    canvas.drawCircle((float) getX(),(float) getY(), 15, paint);
   }
 
   void update(Level level) {
